+++ coin-change.h
@@ -8,8 +8,8 @@ typedef struct {
 extern coinChange* answerStruct;
 extern uint64_t** tabuArray;
 
-int changeRecursive(int amount, int* coins, int coinsSize);
-int change(int amount, int* coins, int coinsSize);
+uint64_t changeRecursive(int amount, int* coins, int coinsSize);
+uint64_t change(int amount, int* coins, int coinsSize);
 coinChange* fillCombArrays(int numberCombinations, int amount, int* coins, int coinsSize);
 
 void free2DArray(uint64_t** tabuArray, int rows);
